--- src/main/resources/META-INF/gradle-plugins/com.f2prateek.checkstyle.properties
@@ -1,1 +1,0 @@
-implementation-class=com.f2prateek.checkstyle.AndroidCheckstylePlugin
