+++ src/main/resources/META-INF/gradle-plugins/com.leinardi.android.checkstyle.properties
@@ -1,0 +1,1 @@
+implementation-class=com.leinardi.android.checkstyle.AndroidCheckstylePlugin
